feat: split `handle_connect` & `bucket` into module
