+++ .vscode/settings.json
@@ -30,6 +30,7 @@
 		"immer",
 		"inbook",
 		"incollection",
+		"indicatif",
 		"inproceedings",
 		"jsonwebtoken",
 		"JWST",
